# AstraServerGo.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AstraServerGo",
    "description": "基于 `AstraScheduleServerGo` 源码生成的 OpenAPI 3.0 文档。\n\n- 共 55 个 API 端点，覆盖客户端课表、天气、WebSocket、管理后台认证、用户管理、学校/年级/班级管理、配置管理、自动任务、倒数日、备份导入导出等功能。\n- JWT 认证接口使用 `jwtAuth` 安全方案（通过 POST /web/auth/login 获取 Token）。\n- 写入接口同时支持 BasicAuth（兼容旧版客户端）和 JWT + 密码验证。\n- 部分接口返回结构是动态 JSON（`map[string]interface{}`），已尽量按实际行为建模。\n",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/api/broadcast/{school}/{grade}/{class_number}": {
      "post": {
        "summary": "广播同步消息（SyncConfig）",
        "description": "向指定级部所有已连接的 WebSocket 客户端发送 SyncConfig 消息，通知客户端拉取最新配置。无需请求体。",
        "operationId": "broadcastSyncConfig",
        "tags": ["websocket"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "class_number", "in": "path", "description": "班级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "广播结果",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BroadcastResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
    "/web/config/{school}/{grade}/subjects/options": {
      "get": {
        "summary": "获取科目下拉选项",
        "operationId": "getSubjectsOptions",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "科目选项列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["options"],
                  "properties": {
                    "options": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": { "type": "string", "description": "显示文本，如「语文（Chinese）」" },
                          "value": { "type": "string", "description": "选项值，如「语文」" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/config/{school}/{grade}/subjects": {
      "get": {
        "summary": "获取科目配置",
        "description": "返回该学校/年级的科目简称与全称映射。",
        "operationId": "getSubjects",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "科目简称与全称",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubjectsPayload" }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新科目配置",
        "description": "支持两种请求体格式：带 `model` 包装（NaiveUI 管理后台格式）或直接传 `abbr`/`fullName`。需要 BasicAuth 认证。",
        "operationId": "putSubjects",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  { "$ref": "#/components/schemas/SubjectsPayload" },
                  {
                    "type": "object",
                    "required": ["model"],
                    "properties": {
                      "model": { "$ref": "#/components/schemas/SubjectsPayload" }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200" }
              }
            }
          },
          "400": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/config/{school}/{grade}/timetable/options": {
      "get": {
        "summary": "获取作息表选项",
        "description": "返回可用的作息表名称列表，每项含 label、value 和该作息所需的课节数（need）。",
        "operationId": "getTimetableOptions",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "作息选项",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["options"],
                  "properties": {
                    "options": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": { "type": "string" },
                          "value": { "type": "string" },
                          "need": { "type": "integer", "description": "该作息表需要的最少课节数" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/config/{school}/{grade}/timetable": {
      "get": {
        "summary": "获取作息配置",
        "description": "返回作息表（timetable）、分隔符（divider）和开学日期（start）。",
        "operationId": "getTimetable",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "作息配置",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TimetableConfig" }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新作息配置",
        "description": "更新作息表配置。必须包含「常日」作息；divider 键会与 timetable 自动同步。需要 BasicAuth 认证。",
        "operationId": "putTimetable",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TimetableConfig" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200" }
              }
            }
          },
          "400": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/config/{school}/{grade}/{class_number}/schedule": {
      "get": {
        "summary": "获取班级课表配置",
        "description": "返回管理端格式的课表配置，classList 为二维数组（每位教师单独一行），课节数不足时自动补「课」。",
        "operationId": "getScheduleConfig",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "class_number", "in": "path", "description": "班级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "课表配置（管理端格式）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["daily_class"],
                  "properties": {
                    "daily_class": {
                      "type": "array",
                      "minItems": 7,
                      "maxItems": 7,
                      "items": { "$ref": "#/components/schemas/DailyClassInput" }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新班级课表配置",
        "description": "支持带 `model` 包装或无包装两种格式。需要 BasicAuth 认证。",
        "operationId": "putScheduleConfig",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "class_number", "in": "path", "description": "班级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "required": ["daily_class"],
                    "properties": {
                      "daily_class": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/DailyClassInput" }
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": ["model"],
                    "properties": {
                      "model": {
                        "type": "object",
                        "required": ["daily_class"],
                        "properties": {
                          "daily_class": {
                            "type": "array",
                            "items": { "$ref": "#/components/schemas/DailyClassInput" }
                          }
                        }
                      }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200" }
              }
            }
          },
          "400": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/config/{school}/{grade}/{class_number}/settings": {
      "get": {
        "summary": "获取班级通用设置",
        "description": "返回倒计时目标、天气预警、星期显示、横幅文本、CSS 样式等通用配置。",
        "operationId": "getSettings",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "class_number", "in": "path", "description": "班级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "通用设置",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClientConfigItems" }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新班级通用设置",
        "description": "需要 BasicAuth 认证。",
        "operationId": "putSettings",
        "tags": ["web-config"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "class_number", "in": "path", "description": "班级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClientConfigItems" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200" }
              }
            }
          },
          "400": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailyClassInput": {
        "description": "一天的课表（管理端格式，classList 为二维数组，每位教师单独一行）",
        "type": "object",
        "properties": {
          "Chinese": { "type": "string" },
          "English": { "type": "string" },
          "classList": {
            "type": "array",
            "items": {
              "type": "array",
              "items": { "type": "string" }
            },
            "description": "每节课教师列表（二维数组）"
          },
          "timetable": { "type": "string" }
        }
      },
...
      "TimetableConfig": {
        "type": "object",
        "properties": {
          "timetable": { "$ref": "#/components/schemas/TimetableMap" },
          "divider": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "integer" }
            },
            "description": "分隔符配置：键为作息名称，值为分隔符位置的课节号数组"
          },
          "start": { "type": "string", "description": "开学日期（YYYY-MM-DD）" }
        }
      },
      "SubjectsPayload": {
        "type": "object",
        "properties": {
          "abbr": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TextItem" },
            "description": "科目简称列表"
          },
          "fullName": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TextItem" },
            "description": "科目全称列表（与 abbr 按索引一一对应）"
          }
        }
      },
      "TextItem": {
        "type": "object",
        "properties": {
          "text": { "type": "string" }
        }
      },
...
      "BroadcastResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "message": { "type": "string", "example": "SyncConfig" },
          "sent": { "type": "integer", "description": "成功发送的 WebSocket 消息数" },
          "websocket_enabled": { "type": "boolean", "description": "WebSocket 功能是否启用" }
        }
      },
...
      "ErrorWithErrorField": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "Status200": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "integer", "example": 200 }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "（已废弃，仅供旧版客户端兼容）用户名：`AstraSchedule` 或 `ElectronClassSchedule`，密码：配置文件 `secret.token`"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，通过 POST /web/auth/login 获取。写入接口还需在 X-Verify-Password 头中提供密码。"
      }
    }
  },
  "servers": [
    {
      "url": "https://{hostvar1}.getastra.cn"
    }
  ],
  "security": []
}